% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_cutter.R
\name{box_cutter}
\alias{box_cutter}
\title{Neuron Box Cutter
The neuron box cutter creates a virtual axis-aligned box with given min and max corners and uses it to "cut" the neuron.
That is, it creates virtual nodes at the intersection of the box with the neuron, adds an empty property with key "cut"
to all branches and nodes that lie outside the reconstruction.
It also adds two properties to the neuron "cutbox_min" and "cutbox_max", the box min and max corner points.
If infinite value is given for certain corner x,y,z value, it takes the neuron min/max value for that axis.}
\usage{
box_cutter(json_info, mins = c(Inf, Inf, Inf), maxs = c(Inf, Inf, Inf))
}
\arguments{
\item{json_info}{info about a reconstruction in JSON format}

\item{mins}{a vector of three values determining the min x,y,z coordinates used to delimit the box. When the value is Inf it takes the min value of that axis}

\item{max}{a vector of three values determining the max x,y,z coordinates used to delimit the box. When the value is Inf it takes the max value of that axis}
}
\value{
a list where the first element is the JSON plain text of the neuron and the second is a list of the elements in the neuron
}
\description{
Neuron Box Cutter
The neuron box cutter creates a virtual axis-aligned box with given min and max corners and uses it to "cut" the neuron.
That is, it creates virtual nodes at the intersection of the box with the neuron, adds an empty property with key "cut"
to all branches and nodes that lie outside the reconstruction.
It also adds two properties to the neuron "cutbox_min" and "cutbox_max", the box min and max corner points.
If infinite value is given for certain corner x,y,z value, it takes the neuron min/max value for that axis.
}
