% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_simulation_models.R
\name{compute_clusters}
\alias{compute_clusters}
\title{Dendrite diamater was estimate by this function. It is consider a latent variable in the model}
\usage{
compute_clusters(data)
}
\arguments{
\item{data}{is a data.frame with the measures computed in the terminal tips}
}
\value{
a data.frame that is the data input param after adding the new diameter column
}
\description{
Dendrite diamater was estimate by this function. It is consider a latent variable in the model
}
