% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cut_nodes.R
\name{get_cut_nodes}
\alias{get_cut_nodes}
\title{Check terminal nodes to select those that have been cut and return the id of that nodes}
\usage{
get_cut_nodes(neuron, n_points_plane = 5)
}
\arguments{
\item{neuron}{is the reconstruction in plain text with JSON format}

\item{n_points_plane}{is an integer denoting the minimum number of points in a plane to be considered a cutting plane}
}
\value{
an array with the id of the terminal nodes that belong to the tips of the cut dendrites
}
\description{
The function computes the 3D convex hull of the terminal nodes. Then search the normal faces that points to the same direction.
If there are several coplanar faces there is a cut and all that terminal nodes are delimiting a cut.
}
