% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_desc_classifier.R
\name{get_features_files}
\alias{get_features_files}
\title{This function compute the features of each node}
\usage{
get_features_files(path, eps = 60)
}
\arguments{
\item{path}{is a string indicating the folder where the files representing neurons are saved}

\item{eps}{is an integer value that represents the level of simplification applied over the neuron}
}
\value{
a dataset with the features of each node. List if features can be found in c_node_feature_extractor
}
\description{
This function compute the features of each node. Thus, for each node defining a neurite a set of features is computed according to node_feature_extractor function
}
\examples{
path<-"/home/universidad/Documents/neuron_data/datos/All"
get_features_files(path,60)
}
