% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_BN_classifier.R
\name{learn_BN}
\alias{learn_BN}
\title{Learn BN}
\usage{
learn_BN(data_learn, wl = FALSE, nump = Inf)
}
\arguments{
\item{data_learn}{a dataset to learn the model}

\item{wl}{is TRUE if we force an arc from the class to all the features and FALSE otherwise}

\item{nump}{an integer. It is the maximum number of parents per node}
}
\value{
a list where its first element is the structure of the BN and the second is the parameters
}
\description{
Learn BN structure and parameters according to the training dataset. There are arcs from class to all the explanatory variables and cannot be in the other way direction.
}
