% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_BN_classifier.R
\name{logSumExp}
\alias{logSumExp}
\title{Compute log sum exp function}
\usage{
logSumExp(x)
}
\arguments{
\item{x}{is a matrix of log likelihood where each column is the log-likelihood of a class}
}
\value{
a vector with the probability of the observations
}
\description{
Compute log sum exp function to avoid underflow
}
