% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Main function. Execute the repairing process}
\usage{
main(file_path, seed = 1)
}
\arguments{
\item{file_path}{is a string representing the path to the .DAT or JSON file containing Neurolucida data}

\item{seed}{is an integer value denoting the seed used to generate random variables}
}
\description{
Main function. Execute the repairing process
}
\examples{
file_path <- "/home/universidad/datos/Cortadas/human cing id if6 porta 2 sec1 cel20.DAT"
file_path <- "/home/universidad/datos/H213III/h213III1.DAT"
repaired_neuron <- main(file_path, 1)
write_2_JSON(repaired_neuron,path="~/datos/repaired_neuron.json")

}
