% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro_converter.R
\name{neuro_converter}
\alias{neuro_converter}
\title{Get reconstruction from a file}
\usage{
neuro_converter(file_path, correct = F, eps = 0)
}
\arguments{
\item{correct}{a boolean value. The converter calls the corect method on each neuron in the reconstruction}

\item{eps}{is a boolean value. Error tolerance for the Ramer-Douglas-Peucker simplification algorithm applied at branch level.}

\item{filepath}{a path to the file from features must be computed}
}
\value{
a data.frame containing the values measured for each neurite in the reconstruction
}
\description{
Given a reconstruction file, the converter guess the format by the file extension and reads it.
Optionally, it tries to correct errors in the reconstruction and applies a simplification algorithm over the branches.
Finally reconstruction is get in R.
}
