% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_feature_extractor.R
\name{node_feature_extractor}
\alias{node_feature_extractor}
\title{Compute node features
The node feature extractor computes a set of prebuilt measures for each node in the reconstruction.
Measure values are not aggregated in any way, they are returned as-is (usually as vectors).
Optionally, it tries to correct errors in the reconstruction.}
\usage{
node_feature_extractor(neuron, omit_apical = F, omit_axon = F,
  omit_dend = F, correct = F, removeZjumps = T, id_nodes = c())
}
\arguments{
\item{neuron}{info about a reconstruction in JSON format}

\item{omit_apical}{a boolean value. If set, apical dendrite is not measured}

\item{omit_axon}{a boolean value. If set, axon is not measured}

\item{omit_dend}{a boolean value. If set, dendrites are not measured}

\item{correct}{a boolean value. The converter calls the correct method on each neuron in the reconstruction}

\item{removeZjumps}{a boolean value. Remove jumps over Z axis produced by the Z jumps in the microscopy imaging}

\item{id_nodes}{is an array of integer with the ids of nodes that represent the tips of the cut dendrites}
}
\value{
a data.frame containing the values measured for each neurite in the reconstruction
}
\description{
Compute node features
The node feature extractor computes a set of prebuilt measures for each node in the reconstruction.
Measure values are not aggregated in any way, they are returned as-is (usually as vectors).
Optionally, it tries to correct errors in the reconstruction.
}
