% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_BN_classifier.R
\name{pred_BN}
\alias{pred_BN}
\title{Predict with BN}
\usage{
pred_BN(BN, pred_data)
}
\arguments{
\item{BN}{is an object of bnlearn containing the BN structure and parameters}

\item{pred_data}{a dataset to predict the classes of the instances}
}
\value{
a data.frame whose columns are the truth class, the probability of each class and the predicted class
}
\description{
Predict the class of a dataset given a BN
}
