% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_bifurcation}
\alias{simulate_bifurcation}
\title{Given a terminal node, simulate the bifurcation of the dendrite}
\usage{
simulate_bifurcation(model, node_info)
}
\arguments{
\item{model}{is a bn model generate by bnlearn}

\item{node_info}{is a data.frame containing the terminal nodes that must grow}
}
\value{
a data.frame with the spherical coordinates of the two new nodes with respect to the terminal node introduced to the function
}
\description{
Given a terminal node, simulate the bifurcation of the dendrite
}
\examples{
path <- "/home/universidad/Documents/neuron_data/datos/All"
data <- get_features_files(path, 60)
data <- compute_clusters(data)
simulation_model <- compute_simulation_models(data)
simulation <- simulate_bifurcation(simulation_model$bifurcation, data[data$node_num_descendant==2,])

}
