% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_continuation}
\alias{simulate_continuation}
\title{Given a terminal node, simulate the grow of the dendrite one step}
\usage{
simulate_continuation(model, node_info)
}
\arguments{
\item{model}{is a bn model generate by bnlearn}

\item{node_info}{is a data.frame containing the terminal nodes that must grow}
}
\value{
a data.frame with the spherical coordinates of the new node with respect to the terminal node introduced to the function
}
\description{
Given a terminal node, simulate the grow of the dendrite one step. A step is a bifurcation or a continuation.
}
\examples{
path <- "/home/universidad/Documents/neuron_data/datos/All"
data <- get_features_files(path, 60)
data <- compute_clusters(data)
simulation_model <- compute_simulation_models(data)
simulation <- simulate_continuation(simulation_model$continue, data[data$node_num_descendant==1])

}
