% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_node_coordinates.R
\name{simulated_node_coordinates}
\alias{simulated_node_coordinates}
\title{Include the simulated nodes into the neuron}
\usage{
simulated_node_coordinates(json_info, simulated_nodes, is_bifurcation, init_id,
  omit_apical = F, omit_axon = F, omit_dend = F, correct = F,
  removeZjumps = T)
}
\arguments{
\item{json_info}{info about a reconstruction in JSON format}

\item{is_bifurcation}{is a vector identifying if the node to simulate is a bifurcation or is a continuation}

\item{init_id}{is an integer denoting the highest id of the neuron, it is used to give ids to the new nodes}

\item{omit_apical}{a boolean value. If set, apical dendrite is not measured}

\item{omit_axon}{a boolean value. If set, axon is not measured}

\item{omit_dend}{a boolean value. If set, dendrites are not measured}

\item{correct}{a boolean value. The converter calls the correct method on each neuron in the reconstruction}

\item{removeZjumps}{a boolean value. Remove jumps over Z axis produced by the Z jumps in the microscopy imaging}

\item{simualted_nodes}{is a matrix with the spherical coordinates of the new nodes with respect to their parent nodes}
}
\value{
a list of three elements, the first one is the neuron in JSON format, the second one is a data.frame representing the neuron information and the third one are the ids given to the simulated nodes
}
\description{
Include the simulated nodes into the neuron representation.
}
