% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_desc_classifier.R
\name{train_num_descendant_model}
\alias{train_num_descendant_model}
\title{This function train a model to distinguish between those nodes that should be terminal, bifurcation or continue the neurite (0, 1 or 2 descendants)}
\usage{
train_num_descendant_model(data)
}
\arguments{
\item{data}{is a dataframe which contains the features of each node}
}
\value{
model is an augmented Bayesian network classifier
}
\description{
This function train a model to distinguish between those nodes that should be terminal, bifurcation or continue the neurite (0, 1 or 2 descendants)
}
\examples{
path<-"/home/universidad/Documents/neuron_data/datos/All"
data<-get_features_files(path,60)
desc_model<-train_num_descendant_model(data)
}
