% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_2_JSON.R
\name{write_2_JSON}
\alias{write_2_JSON}
\title{Write the neuron into a JSON file}
\usage{
write_2_JSON(neuron, file_path = "~/datos/simpl.json")
}
\arguments{
\item{neuron}{info about a reconstruction in JSON format}

\item{file_path}{is a string with the path where the json file must be saved, it must include the name of the file and the extension}
}
\description{
This function gets a neuron an writes it into a JSON file
}
